requirements/requirements.txt
click==7.0
google-cloud-storage==1.36.1